Lab 6
Basically a interactive animation where the user controls the camera. The user will move the camera and press the space bar to take a screenshot of what the camera sees in a ppm format. The camera can be changed from perspective view to orthographic view. This animation demonstrates the concepts of rendering.