Lab 7
A free choice lab. We chose to implement a destructible wall where each piece is a different size when shattered and the wall shatters uniquely each time when the animation is ran. The user controls the camera for this animation. If the user presses the space bar, then a sphere will fire forward. If the sphere hits the wall, then the wall shatters